# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de manipulação de produtos",
        "description": "Essa API tem por objetivo a manipulação de produtos",
        "termsOfService": "http://localhost:3000/terms",
        "contact": {
            "email": "example@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/v1",
            "description": "API de teste"
        }
    ],
    "paths": {
        "/product": {
            "post": {
                "summary": "Cadastro de produto",
                "description": "Rota responsável por cadastrar um novo produto",
                "tags": ["product"],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Mouse USB",
                                        "description": "Mouse USB RGB",
                                        "price": 300
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "User unauthorized"
                    },
                    "400": {
                        "description": "Product already exists"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "type": "object",
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product/findByName": {
            "get": {
                "description": "Busca de um produto pelo nome",
                "summary": "Busca de um produto pelo nome",
                "tags": ["product"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome do produto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "type": "object",
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "jwt"
            }
        }
    }
}